feat: FE 영역 코드 정리
- 코드 정리
- 일부 warning 메시지 제거 대응

Story: #169
