+++ apps/frontend/src/blocks/CreateQuizZone/CandidateQuizzes.tsx
@@ -7,14 +7,15 @@ interface CandidateQuizProps {
 const CandidateQuizzes = ({ quizzes }: CandidateQuizProps) => {
     return (
         <ul>
-            {quizzes.map((quiz) => (
-                <li>
+            {quizzes.map((quiz, i) => (
+                <li key={i}>
                     <div>
                         <span>{quiz.type}</span>
                         <span>{quiz.playTime}</span>
                     </div>
                     <div>{quiz.question}</div>
                     <div>{quiz.answer}</div>
+                    <button onClick={() => removeQuiz(quiz)}>삭제</button>
                 </li>
             ))}
         </ul>
