+++ apps/frontend/src/blocks/CreateQuizZone/CreateQuiz.tsx
@@ -1,24 +1,10 @@
 import { ProblemType, Quiz } from '@/types/quizZone.types.ts';
 import useValidState from '@/hook/useValidInput.ts';
+import { validAnswer, validQuestion, validTime } from '@/utils/validators.ts';
 
 interface CreateQuizProps {
     handleCreateQuiz: (quiz: Quiz) => void;
 }
-
-const validQuestion = (question: string) => {
-    if (question.length <= 0) return '문제를 입력해주세요.';
-    if (question.length > 200) return '문제 길이를 초과하였습니다. 최대 200자';
-};
-
-const validAnswer = (answer: string) => {
-    if (answer.length <= 0) return '답안을 입력해주세요.';
-    if (answer.length > 50) return '답안 길이를 초과하였습니다. 최대 50자';
-};
-
-const validTime = (time: number) => {
-    if (time <= 0) return '제한시간은 0초 보다 커야합니다.';
-    if (time > 60) return '제한시간은 60초를 초과할 수 없습니다.';
-};
 
 const CreateQuiz = ({ handleCreateQuiz }: CreateQuizProps) => {
     const [question, questionInvalidMessage, setQuestion, isInvalidQuestion] =
