+++ apps/frontend/src/blocks/CreateQuizZone/CreateQuizSet.tsx
@@ -1,44 +1,16 @@
 import { useState } from 'react';
-import { Quiz, QuizSet } from '@/types/quizZone.types.ts';
+import { Quiz } from '@/types/quizZone.types.ts';
 import CandidateQuizzes from '@/blocks/CreateQuizZone/CandidateQuizzes.tsx';
 import CreateQuiz from '@/blocks/CreateQuizZone/CreateQuiz.tsx';
 import useValidState from '@/hook/useValidInput.ts';
+import { requestCreateQuizSet } from '@/utils/requests.ts';
+import { validQuizSetName, validQuizzes } from '@/utils/validators.ts';
+import { QUIZ_LIMIT_COUNT } from '@/constants/quiz-set.constants.ts';
 
 interface CreateQuizZoneQuizSetProps {
     handlePrevStepButton?: () => void;
     updateQuizSet?: (quizSetId: string, quizSetName: string) => void;
 }
-
-const QUIZ_LIMIT_COUNT = 10;
-
-const requestCreateQuizSet = async (quizSet: QuizSet) => {
-    const { quizSetName, quizzes } = quizSet;
-
-    const response = await fetch('api/quiz', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({
-            quizSetName,
-            quizDetails: quizzes,
-        }),
-    });
-
-    if (!response.ok) {
-        throw Error();
-    }
-
-    return (await response.json()) as string;
-};
-
-const validQuizSetName = (name: string) => {
-    if (name.length <= 0) return '퀴즈존 이름 입력을 확인하세요';
-    if (name.length > 100) return '퀴즈존 이름 길이를 확인하세요. (최대 100자)';
-};
-
-const validQuizzes = (quizzes: Quiz[]) => {
-    if (quizzes.length === 0) return '퀴즈를 1개 이상 등록해야합니다.';
-    if (quizzes.length > QUIZ_LIMIT_COUNT) return '퀴즈는 최대 10개만 등록할 수 있습니다.';
-};
 
 const CreateQuizSet = ({ handlePrevStepButton, updateQuizSet }: CreateQuizZoneQuizSetProps) => {
     const [name, validNameMessage, setName, isInvalidName] = useValidState<string>(
@@ -67,7 +39,6 @@ const CreateQuizSet = ({ handlePrevStepButton, updateQuizSet }: CreateQuizZoneQu
                 quizzes: quizzes,
             });
 
-            console.log(quizSetId);
             updateQuizSet?.(quizSetId, name);
             handlePrevStepButton?.();
         } catch (error) {
