+++ apps/frontend/src/constants/quiz-set.constants.ts
@@ -1,0 +1,1 @@
+export const QUIZ_LIMIT_COUNT = 10;
